Split utils across DCCs to align with Unreal; Fixed finalizer dependency evaluation
