+++ resource/plugins/python/publisher/collectors/max_scene_publisher_collector.py
@@ -5,7 +5,7 @@ from pymxs import runtime as rt
 
 import ftrack_api
 
-from ftrack_connect_pipeline_3dsmax.utils import custom_commands as max_utils
+from ftrack_connect_pipeline_3dsmax import utils as max_utils
 
 from ftrack_connect_pipeline_3dsmax import plugin
