+++ source/ftrack_connect_pipeline_3dsmax/client/save.py
@@ -2,7 +2,7 @@
 # :copyright: Copyright (c) 2014-2022 ftrack
 
 from ftrack_connect_pipeline_qt.client.save import QtSaveClientWidget
-from ftrack_connect_pipeline_3dsmax.utils import custom_commands as max_utils
+from ftrack_connect_pipeline_3dsmax import utils as max_utils
 
 
 class MaxQtSaveClientWidget(QtSaveClientWidget):
